feat: abstract pong game logic into Game classes
